+++ fuel-tx/src/transaction/types/upgrade.rs
@@ -14,6 +14,7 @@ use crate::{
     Input,
     Output,
     TransactionRepr,
+    TxId,
     ValidityError,
 };
 use educe::Educe;
@@ -271,11 +272,14 @@ impl crate::Cacheable for Upgrade {
         self.metadata.is_some()
     }
 
-    fn precompute(&mut self, chain_id: &ChainId) -> Result<(), ValidityError> {
+    fn precompute(&mut self, chain_id: &ChainId) -> Result<(), (TxId, ValidityError)> {
         self.metadata = None;
+        let common_metadata = CommonMetadata::compute(self, chain_id)?;
+        let body_metadata =
+            UpgradeMetadata::compute(self).map_err(|e| (common_metadata.id, e))?;
         self.metadata = Some(ChargeableMetadata {
-            common: CommonMetadata::compute(self, chain_id)?,
-            body: UpgradeMetadata::compute(self)?,
+            common: common_metadata,
+            body: body_metadata,
         });
         Ok(())
     }