+++ fuel-tx/src/transaction/types/upload.rs
@@ -16,6 +16,7 @@ use crate::{
     Input,
     Output,
     TransactionRepr,
+    TxId,
     ValidityError,
 };
 use core::ops::Deref;
@@ -278,7 +279,7 @@ impl crate::Cacheable for Upload {
         self.metadata.is_some()
     }
 
-    fn precompute(&mut self, chain_id: &ChainId) -> Result<(), ValidityError> {
+    fn precompute(&mut self, chain_id: &ChainId) -> Result<(), (TxId, ValidityError)> {
         self.metadata = None;
         self.metadata = Some(ChargeableMetadata {
             common: CommonMetadata::compute(self, chain_id)?,